+++ src/includes/psmconfig.inc.php
@@ -30,7 +30,7 @@
 /**
  * Current PSM version
  */
-define('PSM_VERSION', '4.2.1-hs');
+define('PSM_VERSION', '4.2.2-hs');
 
 /**
  * URL to check for updates. Will not be checked if turned off on config page.